Main svc events

+++ ewm-main/src/main/java/ru/practicum/ewm/main/service/category/CategoryServiceImpl.java
@@ -9,10 +9,12 @@ import org.springframework.stereotype.Service;
 import ru.practicum.ewm.main.dto.category.CategoryDto;
 import ru.practicum.ewm.main.dto.category.NewCategoryDto;
 import ru.practicum.ewm.main.dto.category.UpdateCategoryDto;
+import ru.practicum.ewm.main.exception.ConflictException;
 import ru.practicum.ewm.main.exception.NotFoundException;
 import ru.practicum.ewm.main.mapper.CategoryMapper;
 import ru.practicum.ewm.main.entity.Category;
 import ru.practicum.ewm.main.repository.CategoryRepository;
+import ru.practicum.ewm.main.repository.EventRepository;
 
 import java.util.List;
 
@@ -20,6 +22,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto createOrUpdate(CategoryDto categoryDto) {
@@ -29,6 +32,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new NotFoundException("Category with id=" + id + " was not found");
+        }
+
+        if (eventRepository.existsByCategoryId(id)) {
+            throw new ConflictException("Cannot delete category with id=" + id + " because it has linked events");
+        }
+
         categoryRepository.deleteById(id);
     }
 
@@ -61,6 +72,16 @@ public class CategoryServiceImpl implements CategoryService {
     public CategoryDto updateCategory(Long id, UpdateCategoryDto updateCategoryDto) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Category with id=" + id + " was not found"));
+
+        if (updateCategoryDto.getName() != null &&
+                !updateCategoryDto.getName().equals(category.getName())) {
+
+            boolean nameExists = categoryRepository.existsByNameAndIdNot(updateCategoryDto.getName(), id);
+            if (nameExists) {
+                throw new ConflictException("Category name already exists");
+            }
+        }
+
         CategoryMapper.updateCategory(category, updateCategoryDto);
         Category updateCategory = categoryRepository.save(category);
         return CategoryMapper.toCategoryDto(updateCategory);
